style: use FormSection for settings section
Refs: CO-1967
