+++ package.json
@@ -87,9 +87,9 @@
 		"@reduxjs/toolkit": "^1.9.7",
 		"@types/webpack": "^5.28.5",
 		"@types/webpack-env": "^1.18.5",
-		"@zextras/carbonio-design-system": "9.0.0",
+		"@zextras/carbonio-design-system": "file:../carbonio-design-system/zextras-carbonio-design-system-0.0.0-semantically-released.tgz",
 		"@zextras/carbonio-shell-ui": "10.0.0",
-		"@zextras/carbonio-ui-preview": "3.1.1",
+		"@zextras/carbonio-ui-preview": "file:../carbonio-ui-preview/zextras-carbonio-ui-preview-0.0.0-semantically-released.tgz",
 		"core-js": "^3.39.0",
 		"date-arithmetic": "^4.1.0",
 		"i18next": "^22.5.1",
