+++ src/settings/components/time-picker.jsx
@@ -5,7 +5,6 @@
  */
 import React from 'react';
 
-import { Container } from '@zextras/carbonio-design-system';
 import { t } from '@zextras/carbonio-shell-ui';
 import momentLocalizer from 'react-widgets-moment';
 
@@ -18,27 +17,23 @@ momentLocalizer();
 export default function DatePicker({ start, end, onChange, day, disabled }) {
 	return (
 		<>
-			<Styler orientation="horizontal" allDay height="fit" mainAlignment="space-between">
-				<Container padding={{ all: 'small' }}>
-					<StartTimePicker
-						disabled={disabled}
-						start={start}
-						showEnd={false}
-						onChange={onChange}
-						day={day}
-						label={t('label.from', 'from')}
-					/>
-				</Container>
-				<Container padding={{ all: 'small' }}>
-					<StartTimePicker
-						label={t('label.to', 'to')}
-						disabled={disabled}
-						start={end}
-						showEnd
-						onChange={onChange}
-						day={day}
-					/>
-				</Container>
+			<Styler orientation="horizontal" allDay height="fit" mainAlignment="flex-start" gap="1rem">
+				<StartTimePicker
+					disabled={disabled}
+					start={start}
+					showEnd={false}
+					onChange={onChange}
+					day={day}
+					label={t('label.from', 'from')}
+				/>
+				<StartTimePicker
+					label={t('label.to', 'to')}
+					disabled={disabled}
+					start={end}
+					showEnd
+					onChange={onChange}
+					day={day}
+				/>
 			</Styler>
 		</>
 	);
